add missing prepareForWrite() calls

+++ idea/src/org/jetbrains/kotlin/idea/inspections/CanBeParameterInspection.kt
@@ -16,6 +16,7 @@
 
 package org.jetbrains.kotlin.idea.inspections
 
+import com.intellij.codeInsight.FileModificationService
 import com.intellij.codeInspection.LocalQuickFix
 import com.intellij.codeInspection.ProblemDescriptor
 import com.intellij.codeInspection.ProblemHighlightType
@@ -124,6 +125,7 @@ class CanBeParameterInspection : AbstractKotlinInspection() {
         override fun getFamilyName() = fix.familyName
 
         override fun applyFix(project: Project, descriptor: ProblemDescriptor) {
+            if (!FileModificationService.getInstance().preparePsiElementForWrite(descriptor.psiElement)) return
             parameter.valOrVarKeyword?.delete()
             // Delete visibility / open / final / lateinit, if any
             // Retain annotations / vararg
