+++ idea/src/org/jetbrains/kotlin/idea/inspections/RedundantSemicolonInspection.kt
@@ -16,6 +16,7 @@
 
 package org.jetbrains.kotlin.idea.inspections
 
+import com.intellij.codeInsight.FileModificationService
 import com.intellij.codeInspection.*
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiComment
@@ -63,6 +64,7 @@ class RedundantSemicolonInspection : AbstractKotlinInspection(), CleanupLocalIns
         override fun getFamilyName() = name
 
         override fun applyFix(project: Project, descriptor: ProblemDescriptor) {
+            if (!FileModificationService.getInstance().preparePsiElementForWrite(descriptor.psiElement)) return
             descriptor.psiElement.delete()
         }
     }
